Base
This is our starting point.

Simulate processing a network request in the request-response cycle in a
way that blocks the page from being rendered immediately.

Note that we're only working with `ActiveModel` instances. There is
nothing stored in the database.

Our objective is to background this, and return the response with Turbo.