Background request and broadcast response
Notable changes include adding a [custom serializer][], and wrapping our
partial with an element identified by [`dom_id`][] so that we can
broadcast to it.

This implementation satisfies our requirements, but its ergonomics could
be improved.

Additionally, the scope for our channel is really broad, since we we're
identifying our element with `#new_order_search`. Ideally, this should
be scoped to the user issuing the request, or to the model instance.

[custom serializer]: https://guides.rubyonrails.org/active_job_basics.html#serializers
[`dom_id`]: https://api.rubyonrails.org/classes/ActionView/RecordIdentifier.html#method-i-dom_id